requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
markdown-it-py==3.0.0
mdurl==0.1.2
modal==1.2.5
multidict==6.7.0
openai==2.11.0
packaging==25.0
postgrest==2.25.1
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.25.1
rich==14.2.0
shellingham==1.5.4
sniffio==1.3.1
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
synchronicity==0.10.5
toml==0.10.2
tqdm==4.67.1
typer==0.20.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
